Set up test harness for fixtures tests with errors
