+++ crates/crochet/tests/fail/multiple_errors.crochet
@@ -1,0 +1,4 @@
+let add = (a, b) => a + b;
+
+add("hello", "world");
+add(true, false);
