+++ crates/crochet/tests/fail/multiple_errors.error.txt
@@ -1,0 +1,7 @@
+CompileError
+│
+╰┬▶ TypeError::UnificationError: "hello", number
+ │  ╰╴ Unification failure
+ │
+ ╰▶ TypeError::UnificationError: true, number
+    ╰╴ Unification failure
