+++ crates/crochet_cli/Cargo.toml
@@ -18,4 +18,3 @@ tree_sitter_crochet = { version = "0.1.0", path = "../tree_sitter_crochet" }
 error-stack = "0.2.4"
 insta = "1.13.0"
 pretty_assertions = "1.2.1"
-testing_macros = "0.2.5"
